Raise minimum dependencies to php, symfony/* and doctrine/dbal packages
This change raises the minimum php version to 8.1, enabling the usage
of these syntax features:

- property type declarations
- constructor property promotions
- readonly class properties

This change also raises the minimum required versions of symfony/*
packages as all supported packages of symfony require php 8.1

This change also drops support for doctrine/dbal:^2 to ease the
transition to doctrine/dbal:^4 in the future.

Last but not least, phpunit/phpunit:^9.5 is required to run tests.